type : components/organisms 폴더에 있는 타입들 분리
